+++ Rocket.Unturned/Player/UnturnedPlayer.cs
@@ -198,7 +198,10 @@ namespace Rocket.Unturned.Player
 
         public void TriggerEffect(ushort effectID)
         {
-            EffectManager.instance.channel.send("tellEffectPoint", CSteamID, ESteamPacket.UPDATE_UNRELIABLE_BUFFER, new object[] { effectID, player.transform.position });
+            TriggerEffectParameters parameters = new TriggerEffectParameters(effectID);
+            parameters.position = player.transform.position;
+            parameters.relevantPlayerID = CSteamID;
+            EffectManager.triggerEffect(parameters);
         }
         
         public string IP
@@ -345,16 +348,9 @@ namespace Rocket.Unturned.Player
 
         public void Teleport(Vector3 position, float rotation)
         {
-            if (VanishMode)
-            {
-                player.channel.send("askTeleport", ESteamCall.OWNER, ESteamPacket.UPDATE_RELIABLE_BUFFER, position, MeasurementTool.angleToByte(rotation));
-                player.channel.send("askTeleport", ESteamCall.NOT_OWNER, ESteamPacket.UPDATE_RELIABLE_BUFFER, new Vector3(position.y, position.y + 1337, position.z), MeasurementTool.angleToByte(rotation));
-                player.channel.send("askTeleport", ESteamCall.SERVER, ESteamPacket.UPDATE_RELIABLE_BUFFER, position, MeasurementTool.angleToByte(rotation));
-            }
-            else
-            {
-				player.teleportToLocation(position, rotation);
-            }
+            // In older versions Rocket had a special case for "vanish" mode, but now the vanilla game will
+            // only send teleport to all clients if canAddSimulationResultsToUpdates is true.
+            player.teleportToLocation(position, rotation);
         }
 
         public bool VanishMode
@@ -466,8 +462,7 @@ namespace Rocket.Unturned.Player
             }
             set
             {
-                player.skills.channel.send("tellExperience", ESteamCall.SERVER, ESteamPacket.UPDATE_RELIABLE_BUFFER, value);
-                player.skills.channel.send("tellExperience", ESteamCall.OWNER, ESteamPacket.UPDATE_RELIABLE_BUFFER, value);
+                player.skills.ServerSetExperience(value);
             }
         }
 
@@ -525,8 +520,7 @@ namespace Rocket.Unturned.Player
             }
             set
             {
-                player.life.tellBroken(Provider.server,value);
-                player.life.channel.send("tellBroken", ESteamCall.OWNER, ESteamPacket.UPDATE_RELIABLE_BUFFER, new object[] { value });
+                player.life.serverSetLegsBroken(value);
             }
         }
         public bool Bleeding
@@ -537,8 +531,7 @@ namespace Rocket.Unturned.Player
             }
             set
             {
-                player.life.tellBleeding(Provider.server, value);
-                player.life.channel.send("tellBleeding", ESteamCall.OWNER, ESteamPacket.UPDATE_RELIABLE_BUFFER, new object[] { value });
+                player.life.serverSetBleeding(value);
             }
         }