+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes should be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 4.9.3.9 - 2021-03-05
+
+### Changed
+- Replaced `tellBleeding` invocation with `serverSetBleeding`.
+- Replaced `tellBroken` invocation with `serverSetLegsBroken`.
+- Replaced `UnturnedPlayer.TriggerEffect` invocation with `EffectManager.triggerEffect`.
+- Replaced `UnturnedEffect` invocation with `EffectManager.triggerEffect`.
+- Replaced `UnturnedPlayer.Experience` invocation with `PlayerSkills.ServerSetExperience`.
+- Vanish mode teleport no longer needs special handling because the vanilla game skips other clients if canAddSimulationResultsToUpdates is false.
+- Updated gesture changed to use value from event.
+
 ## 4.9.3.8 - 2021-02-26
 
 ### Changed